// src/utils/enums/CategoriesEnums.ts
export enum MarketSector {
  Hortifruti = 'Hortifruti',
  Carnes = 'Carnes',
  FriosELaticinios = 'Frios e Laticínios',
  PadariaEConfeitaria = 'Padaria e Confeitaria',
  BebidasNaoAlcoolicas = 'Bebidas não Alcóolicas',
  AlimentosBasicos = 'Alimentos Básicos',
  Congelados = 'Congelados',
  DocesESnacks = 'Doces e Snacks',
  ProdutosDeLimpeza = 'Produtos de Limpeza',
  HigieneEBeleza = 'Higiene e Beleza',
  BebeEInfantil = 'Bebê e Infantil',
  PetShop = 'Pet Shop',
  Mercearia = 'Mercearia',
  BebidasAlcoolicas = 'Bebidas Alcóolicas',
  SaudaveisENaturais = 'Saudáveis e Naturais',
  Outros = 'Outros'
}